@@ -1,9 +1,9 @@
 {
   "name": "jsonc-parser",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "description": "Scanner and parser for JSON with comments.",
   "main": "./lib/umd/main.js",
-  "typings": "./lib/umd/main",
+  "typings": "./lib/umd/main.d.ts",
   "module": "./lib/esm/main.js",
   "author": "Microsoft Corporation",
   "repository": {
@@ -15,13 +15,13 @@
     "url": "https://github.com/microsoft/node-jsonc-parser/issues"
   },
   "devDependencies": {
-    "mocha": "^8.2.1",
-    "typescript": "^4.0.5",
-    "@types/node": "^10.12.12",
-    "@types/mocha": "^5.2.7",
-    "@typescript-eslint/eslint-plugin": "^4.7.0",
-    "@typescript-eslint/parser": "^4.7.0",
-    "eslint": "^7.13.0",
+    "mocha": "^10.0.0",
+    "typescript": "^4.7.4",
+    "@types/node": "^16.x",
+    "@types/mocha": "^9.1.1",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "eslint": "^8.19.0",
     "rimraf": "^3.0.2"
   },
   "scripts": {
