@@ -1,6 +1,6 @@
 {
   "name": "minipass-fetch",
-  "version": "1.4.1",
+  "version": "2.1.2",
   "description": "An implementation of window.fetch in Node.js using Minipass streams",
   "license": "MIT",
   "main": "lib/index.js",
@@ -9,33 +9,42 @@
     "snap": "tap",
     "preversion": "npm test",
     "postversion": "npm publish",
-    "postpublish": "git push origin --follow-tags"
+    "postpublish": "git push origin --follow-tags",
+    "lint": "eslint \"**/*.js\"",
+    "postlint": "template-oss-check",
+    "lintfix": "npm run lint -- --fix",
+    "prepublishOnly": "git push origin --follow-tags",
+    "posttest": "npm run lint",
+    "template-oss-apply": "template-oss-apply --force"
   },
   "tap": {
     "coverage-map": "map.js",
     "check-coverage": true
   },
   "devDependencies": {
-    "@ungap/url-search-params": "^0.1.2",
+    "@npmcli/eslint-config": "^3.0.1",
+    "@npmcli/template-oss": "3.5.0",
+    "@ungap/url-search-params": "^0.2.2",
     "abort-controller": "^3.0.0",
-    "abortcontroller-polyfill": "~1.3.0",
-    "form-data": "^2.5.1",
+    "abortcontroller-polyfill": "~1.7.3",
+    "encoding": "^0.1.13",
+    "form-data": "^4.0.0",
+    "nock": "^13.2.4",
     "parted": "^0.1.1",
     "string-to-arraybuffer": "^1.0.2",
-    "tap": "^15.0.9",
-    "whatwg-url": "^7.0.0"
+    "tap": "^16.0.0"
   },
   "dependencies": {
-    "minipass": "^3.1.0",
+    "minipass": "^3.1.6",
     "minipass-sized": "^1.0.3",
-    "minizlib": "^2.0.0"
+    "minizlib": "^2.1.2"
   },
   "optionalDependencies": {
-    "encoding": "^0.1.12"
+    "encoding": "^0.1.13"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/npm/minipass-fetch.git"
+    "url": "https://github.com/npm/minipass-fetch.git"
   },
   "keywords": [
     "fetch",
@@ -44,10 +53,15 @@
     "window.fetch"
   ],
   "files": [
-    "index.js",
-    "lib/*.js"
+    "bin/",
+    "lib/"
   ],
   "engines": {
-    "node": ">=8"
+    "node": "^12.13.0 || ^14.15.0 || >=16.0.0"
+  },
+  "author": "GitHub Inc.",
+  "templateOSS": {
+    "//@npmcli/template-oss": "This file is partially managed by @npmcli/template-oss. Edits may be overwritten.",
+    "version": "3.5.0"
   }
 }
