@@ -11,7 +11,7 @@
     "bindings",
     "gyp"
   ],
-  "version": "8.4.1",
+  "version": "9.3.0",
   "installVersion": 9,
   "author": "Nathan Rajlich <nathan@tootallnate.net> (http://tootallnate.net)",
   "repository": {
@@ -25,8 +25,8 @@
     "env-paths": "^2.2.0",
     "glob": "^7.1.4",
     "graceful-fs": "^4.2.6",
-    "make-fetch-happen": "^9.1.0",
-    "nopt": "^5.0.0",
+    "make-fetch-happen": "^10.0.3",
+    "nopt": "^6.0.0",
     "npmlog": "^6.0.0",
     "rimraf": "^3.0.2",
     "semver": "^7.3.5",
@@ -34,7 +34,7 @@
     "which": "^2.0.2"
   },
   "engines": {
-    "node": ">= 10.12.0"
+    "node": "^12.22 || ^14.13 || >=16"
   },
   "devDependencies": {
     "bindings": "^1.5.0",
@@ -45,6 +45,6 @@
   },
   "scripts": {
     "lint": "standard */*.js test/**/*.js",
-    "test": "npm run lint && tap --timeout=120 test/test-*"
+    "test": "npm run lint && tap --timeout=600 test/test-*"
   }
 }
