@@ -1,14 +1,20 @@
 {
   "name": "nopt",
-  "version": "5.0.0",
+  "version": "6.0.0",
   "description": "Option parsing for Node, supporting types, shorthands, etc. Used by npm.",
-  "author": "Isaac Z. Schlueter <i@izs.me> (http://blog.izs.me/)",
+  "author": "GitHub Inc.",
   "main": "lib/nopt.js",
   "scripts": {
     "preversion": "npm test",
     "postversion": "npm publish",
     "prepublishOnly": "git push origin --follow-tags",
-    "test": "tap test/*.js"
+    "test": "tap",
+    "lint": "eslint \"**/*.js\"",
+    "postlint": "template-oss-check",
+    "template-oss-apply": "template-oss-apply --force",
+    "lintfix": "npm run lint -- --fix",
+    "snap": "tap",
+    "posttest": "npm run lint"
   },
   "repository": {
     "type": "git",
@@ -19,16 +25,29 @@
   },
   "license": "ISC",
   "dependencies": {
-    "abbrev": "1"
+    "abbrev": "^1.0.0"
   },
   "devDependencies": {
-    "tap": "^14.10.6"
+    "@npmcli/eslint-config": "^3.0.1",
+    "@npmcli/template-oss": "3.5.0",
+    "tap": "^16.3.0"
+  },
+  "tap": {
+    "lines": 87,
+    "functions": 91,
+    "branches": 81,
+    "statements": 87
   },
   "files": [
-    "bin",
-    "lib"
+    "bin/",
+    "lib/"
   ],
   "engines": {
-    "node": ">=6"
+    "node": "^12.13.0 || ^14.15.0 || >=16.0.0"
+  },
+  "templateOSS": {
+    "//@npmcli/template-oss": "This file is partially managed by @npmcli/template-oss. Edits may be overwritten.",
+    "windowsCI": false,
+    "version": "3.5.0"
   }
 }
