@@ -1,7 +1,7 @@
 {
 	"name": "resolve",
 	"description": "resolve like require.resolve() on behalf of files asynchronously and synchronously",
-	"version": "1.22.0",
+	"version": "1.22.1",
 	"repository": {
 		"type": "git",
 		"url": "git://github.com/browserify/resolve.git"
@@ -17,6 +17,7 @@
 		"module"
 	],
 	"scripts": {
+		"prepack": "npmignore --auto --commentLines=autogenerated",
 		"prepublishOnly": "safe-publish-latest && cp node_modules/is-core-module/core.json ./lib/ ||:",
 		"prepublish": "not-in-publish || npm run prepublishOnly",
 		"prelint": "eclint check $(git ls-files | xargs find 2> /dev/null | grep -vE 'node_modules|\\.git')",
@@ -29,20 +30,22 @@
 		"test:multirepo": "cd ./test/resolver/multirepo && npm install && npm test"
 	},
 	"devDependencies": {
-		"@ljharb/eslint-config": "^20.2.0",
+		"@ljharb/eslint-config": "^21.0.0",
 		"array.prototype.map": "^1.0.4",
 		"aud": "^2.0.0",
 		"copy-dir": "^1.3.0",
 		"eclint": "^2.8.1",
-		"eslint": "^8.7.0",
+		"eslint": "=8.8.0",
 		"in-publish": "^2.0.1",
 		"mkdirp": "^0.5.5",
 		"mv": "^2.1.1",
+		"npmignore": "^0.3.0",
 		"object-keys": "^1.1.1",
 		"rimraf": "^2.7.1",
 		"safe-publish-latest": "^2.0.0",
+		"semver": "^6.3.0",
 		"tap": "0.4.13",
-		"tape": "^5.4.1",
+		"tape": "^5.5.3",
 		"tmp": "^0.0.31"
 	},
 	"license": "MIT",
@@ -55,8 +58,14 @@
 		"url": "https://github.com/sponsors/ljharb"
 	},
 	"dependencies": {
-		"is-core-module": "^2.8.1",
+		"is-core-module": "^2.9.0",
 		"path-parse": "^1.0.7",
 		"supports-preserve-symlinks-flag": "^1.0.0"
+	},
+	"publishConfig": {
+		"ignore": [
+			".github/workflows",
+			"appveyor.yml"
+		]
 	}
 }
