@@ -1,7 +1,7 @@
 {
   "name": "source-map",
   "description": "Generates and consumes source maps",
-  "version": "0.7.3",
+  "version": "0.7.4",
   "homepage": "https://github.com/mozilla/source-map",
   "author": "Nick Fitzgerald <nfitzgerald@mozilla.com>",
   "contributors": [
@@ -75,13 +75,14 @@
     "toc": "doctoc --title '## Table of Contents' README.md && doctoc --title '## Table of Contents' CONTRIBUTING.md"
   },
   "devDependencies": {
-    "doctoc": "^0.15.0",
+    "doctoc": "^1.3.1",
     "eslint": "^4.19.1",
     "live-server": "^1.2.0",
     "npm-run-all": "^4.1.2",
     "nyc": "^11.7.1",
     "watch": "^1.0.2",
-    "webpack": "^3.10"
+    "webpack": "^4.9.1",
+    "webpack-cli": "^3.1"
   },
   "nyc": {
     "reporter": "html"
