@@ -1,6 +1,7 @@
 import { ProductService } from './../product.service';
 import { Product } from './../product.module';
 import { Component, OnInit } from '@angular/core';
+import { catchError, Observable, of } from 'rxjs';
 
 @Component({
   selector: 'app-product-read',
@@ -9,10 +10,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ProductReadComponent implements OnInit {
 
-  products!: Product[]
+  products$: Observable<Product[]>;
+
+  products: Product[] = [];
   displayedColumns = ['id', 'name', 'reference', 'price', 'action']
 
-  constructor(private productService: ProductService) { }
+  constructor(private productService: ProductService) {
+    this.products$ = this.productService.read().pipe(
+      catchError(error => {
+        return of([])
+      })
+    );
+  }
 
   ngOnInit(): void {
     this.productService.read().subscribe(products => {
